Cart logic finished. Prepared Flask CRUD model.
