+++ dishes/blueprint.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+
+dishes = Blueprint('dishes', __name__, template_folder='templates')
