+++ templates/index.html
@@ -1,0 +1,9 @@
+{% extends 'base.html' %}
+
+{% block title %}
+    Index page
+{% endblock %}
+
+{% block content %}
+    <h1>Hello from Index.html</h1>
+{% endblock %}
